pyproject.toml
[project]
name = "vlnr"
version = "0.1.0"
description = "Vulnerability-aware Python Project Finder"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "packaging>=23.0",
    "rich>=13.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.3",
    "openai>=2.32.0",
    "instructor>=1.15.1",
    "python-dotenv>=1.2.2",
]

[project.scripts]
poc-find-candidates = "vlnr.cli:app"
poc-scan-vulnerabilities = "vlnr.vuln_cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.9",
    "types-pyyaml>=6.0.12.20260408",
    "vcrpy>=8.1.1",
]

[tool.setuptools]
packages = ["vlnr"]

[tool.uv]
package = true

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.mypy]
strict = true
ignore_missing_imports = true
